remove trailing whitespace
I suspect that changes made directly on GitHub added trailing whitespace 
to some lines. This whitespace should be removed, otherwise when changes 
are made with text editors, this whitespace is automatically removed due 
to the editorconfig. Removing this whitespace would thus become part of 
other commits, hiding the actual changes to the protocol.